add weather day facade

+++ weatherapp-api/src/main/java/me/marcelberger/weatherapp/api/data/weather/WeatherSummaryData.java
@@ -10,7 +10,7 @@ public class WeatherSummaryData {
     private Double temperatureAvg;
     private Double temperatureMax;
     private Double temperatureMin;
-    private Integer humidityAvg;
+    private Double humidityAvg;
     private Integer humidityMax;
     private Integer humidityMin;
     private Double rainTotal;
